backend/event-writer-service/requirements.txt
Flask==2.3.2
flask_cors
kafka-python==2.0.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
